[TASK] Refactor Backend Live Search functionality
The backend search (aka Live Search) can search for database
records only, currently.

To make the search able to find anything other than database records a
`SearchProviderInterface` is introduced which allows to add new providers
in the future and may be used by extension authors in the long run.

As the first implementations, the current database record search is
extracted into a `DatabaseRecordProvider`, which excludes records from
the `pages` table. Additionally, a new `PageRecordProvider` is added
which comes with a custom web component that renders additional
information for pages only.

The changes introduced in this patch are not final and may change in
further iterations, therefore all interfaces are declared as "internal"
and documentation is missing on purpose.

Resolves: #92008
Releases: main
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76118
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>

+++ typo3/sysext/backend/Classes/Search/LiveSearch/QueryParser.php
@@ -18,6 +18,7 @@ namespace TYPO3\CMS\Backend\Search\LiveSearch;
 /**
  * Class for parsing query parameters in backend live search.
  * Detects searches for #pages:23 or #content:mycontent
+ *
  * @internal This class is a specific Backend controller implementation and is not considered part of the Public TYPO3 API.
  */
 class QueryParser
@@ -88,7 +89,7 @@ class QueryParser
      */
     public function getId($query)
     {
-        return str_replace(self::COMMAND_KEY_INDICATOR, '', $query);
+        return (int)str_replace(self::COMMAND_KEY_INDICATOR, '', $query);
     }
 
     /**
